pyproject.toml
[project]
name = "streamlit-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.4",
    "strands-agents>=1.9.1",
    "streamlit>=1.49.1",
]
